extract shared functions
